fix(desktop): resolve Copilot review findings on PR #98 (issue #62)
- Drop the StubEngine import from backend/index.ts: dead since the
  node-mode default became resolveNodeEngine (Copilot thread on
  backend/index.ts).
- Move ModelNotConfiguredError from inference/llama-engine.ts into
  backend/types.ts (the EngineSurface contract module) so the transport
  maps the 503 without importing a concrete engine implementation
  (Copilot thread on server.ts). llama-engine.ts re-exports it for the
  established import path; runtime behavior identical.

Suite 27 files / 196 tests green; compile clean; frozen checks
re-verified (C1 needed settle-retries for the known httpx 5s cold-load
flake; probe: cold /ask 200 in 2580ms; logs C*.postpr3-5 retained).
